Add add_view_generation function in CMake
This function adds allows to target to execute the "genere" tool to
generate the sources *View.h/cpp from the ".dd" GUI specifications
files. The function replaces the old "genereall.cmd" scripts.

The "genere" tool is called by means of a python script to portably save
the logs into files.

Finally, the .dd files may be used in multiple sub-projects. The
add_view_generation function were usually called in the directory
containing the  ".dd" file. Instead they are now called in the directory
where the .cpp/.h files are needed.