Generate wiews from dd files in cmake
Add generates_gui module in CMake. Its purpose is to replace the files
genereall.cmd

This cmake function generates_gui is redrirected to a portable python script. This script calls the genere
binary who produces *View.h/cpp starting from .dd files.

The .dd files may be used in multiple projects. The genere_all.cmd were usually called in the .dd directory.
generate_gui is called where the view .cpp/h files are used. In this way it exposes the dependency between
the target library and the dd file.

While the dd.log file can be considered as source file (the cmake projects depend on it
for the genere_views command), we prefer not to track them on git. Pro is
obviously that we don't clone/pull log files, cons is the sources and dd.log files may be
not synchronized and may cause some issue int the generate_views process.