Fix CMake not building the JARs properly
The khiops.jar contains only icons/images. They were in the "empty"
namespace but the CMake add_jar directive does not work well when the
namespace is empty.  The solution is to always use a namespace for the jar images.

Additionally the add_jar code was refactored to add_khiops_jar function
in the top CMakeLists.txt