Add release job to packaging CICD workflows
We add a release job that is only executed when the workflow executed on
a tag (either by pushing or by manual execution). The workflow creates
the release as a pre-release so it does not disturb the current stable
one in the github page. The workflow may be executed many times as long
as the release stays in the "pre-release" state.

In the release step, we rename the package file names so their version
match the tag. For now, the maintainers must check the coherence of the
"source version" (that in KWKhiopsVersion.cpp) when creating a release.

The release step depends only ond the build step and the tests steps
are now allowed to fail. It is to maintainers to clean any defectiva
package from a final release.

Other opportunistic changes in this commit:
- Update to version 4 of actions checkout, upload-artifact and
  download-artifact. This is to avoid deprecation warnings and in the
  case of dowload-artifact it even necessary as we use the new "pattern"
  option.
- Simplify names of workflows and jobs so they are better seen in the
  "checks" box and tab of a PR. The verbose steps names are kept.  This
  is to achieve a good compromise between legibility in the GitHub webUI
  and auto-documentation.