Fix jvm dependency lib not being found in Windows
When launching Khiops with GUI in windows the `jvm.dll` library is
loaded.  This library has as dependency `vcruntime140.dll` (Visual C
Runtime) which is usually installed by default in Windows.

In case it is not installed Khiops failed without any meaningful
message. To fix this, in the launcher script:
- We add %JAVA_HOME%\bin to the path so `vcruntime140.dll` is always
  found (JRE bundles it).
- We add a new error window that is launched when Khiops exits with
  return code different from 0 o 2, showing the log.