Fix clang-format formatting version macros
The version macros definitions were being "fixed" by clang-format
because they contained a '-' which was interpreted as a minus sign.

To fix it we add an exception to the WhitespaceSensitiveMacros list in
the clang-format config. We change the macro `str` to `KHIOPS_STR` to
avoid potential collisions.

Also we add to CMake the KHIOPS_VERSION variable which is set to the new
versioning scheme. Otherwise cmake refuses to set it as a project's
version.

(cherry picked from commit df6f20e5a1ec3a8eb2c97e928f3abb8cff64d748)