force gcc optimization to O1 in KWDGPODiscretizer
With gcc 12.2.0 (on debian 12), gcc seems too agressive. The only way to fix a segmentation fault
is to force the optimization level to O1 instead of O2.

The segmentation fault occured while accessing the variable cell1 which is NULL. By adding the
following (unnecessary) line, the segmentation fault vanishes:
if (cell1==NULL or cell2==NULL) exit(1);