remove post-installations script from khiops packages (#245)
* remove post-installations script from khiops packages

Java configuration is in khiops-env in the package khiops-core. It is not updated by khiops package anymore.
The trick is that this configuration is triggered by the presence of khiops.jar

On the docker files, We pin the jdk version to the smallest one on each distro (it matches most of the time the default-jdk package)
The idea is to build norm.jar with the oldest jdk to have a jar compatible with all jre. Ideally we should build with jdk 1.8 on all
distros but this package is not available everywhere.