Add Java Temurin to windows installer
Use Eclipse Temurin JRE in the windows installer. The installer is no longer installed on the system while the Khiops installation. It is simply copyed on KHIOPS_HOME.
We're no longer dependent on (de)installation and updates of java on the system. And there is no need to look into the windows register.