Add khiops_env in conda packages
With 3 new environment variables KHIOPS_PATH, KHIOPS_COCLUSTERING_PATH, KHIOPS_ENV_PATH, the scripts khiops,
khiops_coclustering and khiops-env are generic, we can adapt them to all plateforms. These variables are set
in the 3 scripts by cmake at the configuration step.

The khiops-core packages contains 3 more files: khiops_env, khiops and khiops-coclustering (these 2 lasts
are not mandatory but it facilitates the debugging and tests)

The MODL binary is used to get proc number in khiops_env

On UNIXes we have khiops-env, on windows khiops_env.cmd. Renaming of khiops-env to khiops_env to uniformize names between Windows and UNIXes.

Khiops_env for UNIXes
- Add function get_script_dir funtion in khiops-env.
in conda environment, the PREFIX variable is not always set. In order to set the binaries path, we use this new function.
The khiops binaries are collocated with the khiops-env script (except on fedora-like os)

NSIS
- Add replaceInfile function in NSIS, used to set the KHIOPS_HOME variable during the installation.
- NSIS code simplification

Improve khiops_env for windows
- get KHIOPS_PROC_NUMBER with command line
- add a new variable GUI which is set by cmake or nsis. This variable conditions the setting of the java related variables.
- add KHIOPS_MPI_ERROR, KHIOPS_JAVA_ERROR and KHIOPS_COCLUSTERING_PATH to fit to the UNIXes script
- We use %~dp0 to set KHIOPS_HOME, it is the directory of the script. It is no longer necessary to set KHIOPS_HOME with NSIS (or cmake) or export it.
- check mpi implementation of mpiexec

Improve khiops.cmd
- use %~dp0 to locate khiop_env.cmd (no need for KHIOPS_HOME anymore)
- use KHIOPS_MPI_ERROR and KHIOPS_JAVA_ERROR

install.cmake
- produce khiops_env.cmd
- minor fixes in UNIXes scripts
- Improve KNI package for conda: put the .dll in the bin directory and the .lib in the lib directory (as it is for KNI.zip).
- Call khiops_env -env in action test-khiops-on-iris

conda.yml
- use cmd shell for testing conda package on windows
- force process number to test in parallel

Modify khiops_env.cmd API to be the same on Windows and UNIXes
- When using the '--env' flag, variables will only be displayed without changing the environment. It is useful in debugging.

Improve the action test-khiops-install
- test khiops-env playing with KHIOPS_PROC_NUMBER