Don't add $CONDA_PREFIX/bin to the PATH on Linux and MacOS
Instead, add the path to Conda environment bin directory to the
environment, *as a specific environment variable*.

Thus, we match "Conda-based" environments as much as possible.

A "Conda-based" environment is a Conda environment whose packages are
launched outside of the environment, as e.g. when using the Khiops Python
library from a JuPyter notebook which:
- is installed outside the Conda environment where Khiops is installed
- uses a Python kernel (via the `ipykernel` Conda package) which is
  installed inside the Conda envirovnment.

This configuration results in two particularities:
- `sys.path` is correctly set according to the Conda environment (as
  the Python kernel is installed therein);
- `PATH` does not contain `$CONDA_PREFIX/bin`.