Minor bug fixes in parallel library
* Start and stop task progression when aggregate after errors

In parallel tasks, when an error occurs, the master send the order to stop slaves.
The master may receive messages already sent and then it is possible to launch the aggregation.
In that case we have to encapsulate the aggregate process in a progression task in order to pass the assertions

* In KWFileIndexerTask::MasterFinalize

Move an assertion in the if (bProcessEndedCorrectly) block.
It was wrongly triggered in a user interruption/