Fix GetSortedLineNumber() in KWChunkSorterTask
Line number is computed during the sort in the SlaveProcess. Singleton chunks, are not processed by SlaveProcess unless the separators are different.
Therefore, lines are not counted. We now increment the lines of the singleton in the MasterPrepareInput.

Lines number is now aggregated in KWSortedChunkBuilderTask the built buckets contains the size of the bucket and the line number
Minor refactoring in KWSortedChunkBuilderTask, we no longer use an ObjectDictionary to store and access buckets,
instead we access the buckets directly from the shared object array.