signature of the windows installer (#720)
We use a github action from digicert to install smctl utility.
We sign all binaries (as well as the dll and the installer).
We use a test signature with a limited lifespan and a production signature.
The production signature is only used for releases based on tags that are neither release candidate nor alphas