Improve user interruption detection in the Parallel Library
In case of user interruption, the master receives the end of the slave initialisation.
If the slave initialisation fails, the state of the slave is set to 'ERROR' and no other message is expected from this slave.

TaskProgression is set to unterrupted when PLMPISlaveProgressionManger detects an interruption.
This removes unnecessary calls to the timer and to PLMPISlaveProgressionManger.

Remove unused method PLMPISlave::WaitForMessage()